Add the instance action items menu to the immersive editor
