+++ frontend/src/mixins/Instance.js
@@ -6,6 +6,7 @@ import SnapshotApi from '../api/projectSnapshots.js'
 
 import alerts from '../services/alerts.js'
 import Dialog from '../services/dialog.js'
+import { InstanceStateMutator } from '../utils/InstanceStateMutator.js'
 
 import permissionsMixin from './Permissions.js'
 
@@ -13,12 +14,49 @@ export default {
     mixins: [permissionsMixin],
     computed: {
         ...mapState('account', ['teamMembership', 'team']),
+        actionsDropdownOptions () {
+            const flowActionsDisabled = !(this.instance.meta && this.instance.meta.state !== 'suspended')
+
+            const instanceStateChanging = this.instance.pendingStateChange || this.instance.optimisticStateChange
+
+            const result = [
+                {
+                    name: 'Start',
+                    action: this.startInstance,
+                    disabled: instanceStateChanging || this.instanceRunning
+                },
+                { name: 'Restart', action: this.restartInstance, disabled: instanceStateChanging || flowActionsDisabled },
+                { name: 'Suspend', class: ['text-red-700'], action: this.showConfirmSuspendDialog, disabled: instanceStateChanging || flowActionsDisabled }
+            ]
+
+            if (this.hasPermission('project:delete')) {
+                result.push(null)
+                result.push({ name: 'Delete', class: ['text-red-700'], action: this.showConfirmDeleteDialog })
+            }
+
+            return result
+        },
+        instanceRunning () {
+            return this.instance?.meta?.state === 'running'
+        },
         isVisitingAdmin: function () {
             return this.teamMembership?.role === Roles.Admin
         },
         isHA () {
             return this.instance?.ha?.replicas !== undefined
         }
+    },
+    data () {
+        return {
+            instance: {},
+            loading: {
+                deleting: false,
+                suspend: false
+            }
+        }
+    },
+    watch: {
+        instance: 'instanceChanged'
     },
     methods: {
         showConfirmDeleteDialog () {
@@ -66,6 +104,58 @@ export default {
                     hash: this.$router.currentRoute.value.hash
                 })
             }
+        },
+        instanceUpdated (newData) {
+            this.instanceStateMutator.clearState()
+            this.instance = { ...this.instance, ...newData }
+        },
+        instanceChanged () {
+            this.instanceStateMutator = new InstanceStateMutator(this.instance)
+        },
+        async startInstance () {
+            this.instanceStateMutator.setStateOptimistically('starting')
+
+            try {
+                await InstanceApi.startInstance(this.instance)
+                this.instanceStateMutator.setStateAsPendingFromServer()
+            } catch (err) {
+                console.warn('Instance start failed.', err)
+                alerts.emit('Instance start failed.', 'warning')
+                this.instanceStateMutator.restoreState()
+            }
+        },
+        async restartInstance () {
+            this.instanceStateMutator.setStateOptimistically('restarting')
+            try {
+                await InstanceApi.restartInstance(this.instance)
+                this.instanceStateMutator.setStateAsPendingFromServer()
+            } catch (err) {
+                console.warn('Instance restart failed.', err)
+                alerts.emit('Instance restart failed.', 'warning')
+                this.instanceStateMutator.restoreState()
+            }
+        },
+        deleteInstance () {
+            const applicationId = this.instance.application.id
+            this.loading.deleting = true
+            InstanceApi.deleteInstance(this.instance)
+                .then(() => this.$router.push({ name: 'ApplicationInstances', params: { id: applicationId } }))
+                .then(() => alerts.emit('Instance successfully deleted.', 'confirmation'))
+                .catch(err => {
+                    console.warn(err)
+                    alerts.emit('Instance failed to delete.', 'warning')
+                    this.loading.deleting = false
+                })
         }
+    },
+    async created () {
+        await this.loadInstance()
+
+        this.$watch(
+            () => this.$route.params.id,
+            async () => {
+                await this.loadInstance()
+            }
+        )
     }
 }
